+++ src/main/java/br/com/ucs/MusicHarmony/repository/QuestionsRepository.java
@@ -4,6 +4,7 @@ import br.com.ucs.MusicHarmony.model.Questions;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+// Parecida com a lógica do DAO, serve para mapear a consulta e persistir os dados.
 @Repository
-public interface QuestionsRepository extends JpaRepository<Questions, Integer> {
+public interface QuestionsRepository extends JpaRepository<Questions, Long> {
 }
